Migrate data sinks in notifications-jvm module (#1620)
