--- streampipes-extensions/streampipes-sinks-notifications-jvm/src/main/java/org/apache/streampipes/sinks/notifications/jvm/onesignal/OneSignalParameters.java
@@ -1,49 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package org.apache.streampipes.sinks.notifications.jvm.onesignal;
-
-import org.apache.streampipes.model.graph.DataSinkInvocation;
-import org.apache.streampipes.wrapper.params.binding.EventSinkBindingParams;
-
-public class OneSignalParameters extends EventSinkBindingParams {
-
-  private String title;
-  private String content;
-  private String appId;
-  private String apiKey;
-
-  public OneSignalParameters(DataSinkInvocation graph, String content, String appId, String apiKey) {
-    super(graph);
-    this.content = content;
-    this.appId = appId;
-    this.apiKey = apiKey;
-  }
-
-  public String getContent() {
-    return content;
-  }
-
-  public String getAppId() {
-    return appId;
-  }
-
-  public String getApiKey() {
-    return apiKey;
-  }
-}
